requirements.txt
asgiref==3.7.2
Django==4.2.4
django-autoslug==1.9.9
django-dirtyfields==1.9.2
django-phonenumber-field==7.1.0
django-simple-history==3.3.0
django-versatileimagefield==3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
phonenumbers==8.13.19
Pillow==10.0.0
PyJWT==2.8.0
python-magic==0.4.27
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
